# perf/perf_duplicate: Minor testcase fixes