perf/perf_duplicate: Fix code style warnings
    Rename output variable to fix code style warnings.
    Shorten the failure message as well to avoid code style warning.    
perf/perf_duplicate: Fix wrong package dependency
    Testcase has a wrong package dependecy on kernel-default-base-debuginfo.
    Change it to kernel-default-debuginfo.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>