@sacsant 2022-04-27
Before Fix:

avocado run --test-runner runner perf/perf_duplicate_probe.py 
JOB ID     : 641636720258cfb9f5c6ff3b5ec3ed00de7d18a7
JOB LOG    : /home/avocado/results/job-2022-04-27T01.41-6416367/job.log
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: CANCEL: kernel-default-base-debuginfo is needed for the test to be run (0.45 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 1
JOB HTML   : /home/avocado/results/job-2022-04-27T01.41-6416367/results.html
JOB TIME   : 8.45 s

After fix:

avocado run --test-runner runner perf/perf_duplicate_probe.py 
JOB ID     : 376ed9179705dea71076bd97317b4b8dd4c1cd98
JOB LOG    : /home/avocado/results/job-2022-04-27T01.53-376ed91/job.log
 (1/1) perf/perf_duplicate_probe.py:PerfProbe.test_probe: PASS (1.48 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado/results/job-2022-04-27T01.53-376ed91/results.html
JOB TIME   : 9.50 s